// src/app/interfaces/book.interface.ts
import { Model } from "mongoose";

export interface IBook {
  title: string;
  author: string;
  genre:
    | "FICTION"
    | "NON_FICTION"
    | "SCIENCE"
    | "HISTORY"
    | "BIOGRAPHY"
    | "FANTASY";
  isbn: string;
  description: string;
  copies: number;
  available: boolean;
}

export interface IBookStaticMethod extends Model<IBook> {
  updateAvailable(bookId: string, quantity: number): Promise<void>;
}